Move configuration file support to main utilities module
As clients may also be interested in getting settings like the publisher
endpoint.

As a result, TOML support has been included among main dependencies, however
it just uses the standard library with Python >= 3.11 and a pure Python
implementation otherwise.
