+++ pyproject.toml
@@ -36,6 +36,7 @@ dependencies = [
     "httpx",
     "numpy",
     "pytest",
+    "tomli>=2;python_version<\"3.11\"",
 ]
 
 [tool.hatch.version]
@@ -48,7 +49,6 @@ services = [
     "fastapi",
     "fastapi_websocket_pubsub",
     "safer",
-    "tomli>=2;python_version<\"3.11\"",
     "uvicorn",
     "watchfiles",
 ]
